+++ FlyingAces2/Assets/Scripts/UI Scripts/TutorialHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aefdab917a87a304b8cbb47f4ccc4a1b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
