+++ FlyingAces2/Assets/Scripts/UI Scripts/UIHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9d0c860664a77544b0290ca4d820bac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
